+++ packages/reactified/index.ts
@@ -1,3 +1,3 @@
-export * from './functions'
 export * from './builtins'
 export * from './math'
+export * from './string'
